Dependency definitions refactored.
